inventory character and weapon system (v1)

+++ Counter Blood Quest/Assets/Scripts/SO/Character/CharacterDefaultData.cs
@@ -13,7 +13,7 @@ namespace SO.Character
         [SerializeField] private int defaultArmor = 0;
 
         [Header("Изначальное значение скорости передвижения персонажа")]
-        [SerializeField] private int defaultSpeedMovement = 5;
+        [SerializeField] private float defaultSpeedMovement = 5;
 
         [Header("Изначальное значение скорости вращения персонажа")]
         [SerializeField] private int defaultSpeedRotating = 10;
@@ -22,7 +22,7 @@ namespace SO.Character
         #region Properties
         public int DefaultHealth { get => defaultHealth; }
         public int DefaultArmor { get => defaultArmor; }
-        public int DefaultSpeedMovement { get => defaultSpeedMovement; }
+        public float DefaultSpeedMovement { get => defaultSpeedMovement; }
         public int DefaultSpeedRotating { get => defaultSpeedRotating; }
 
         #endregion
